Add optimization of the cost function using SciPy functions
I modified the costFunction and gradientDescent arguments. Now they don't get the prediction using a given theta as an argument. Now they get theta and X as arguments and calculate the prediction calling to hypothesis. This allows for a straightforward use of the optimization function.

Also added comments to the body.